unify nfo parsing
