+++ src/sonarr_metadata_rewrite/metadata_processor.py
@@ -7,6 +7,7 @@ from xml.etree.ElementTree import ElementTree  # noqa: F401
 
 from sonarr_metadata_rewrite.backup_utils import create_backup, get_backup_path
 from sonarr_metadata_rewrite.config import Settings
+from sonarr_metadata_rewrite.file_utils import extract_metadata_info
 from sonarr_metadata_rewrite.models import (
     MetadataInfo,
     MetadataProcessResult,
@@ -14,7 +15,6 @@ from sonarr_metadata_rewrite.models import (
     TranslatedContent,
     TranslatedString,
 )
-from sonarr_metadata_rewrite.retry_utils import retry
 from sonarr_metadata_rewrite.translator import Translator
 
 logger = logging.getLogger(__name__)
@@ -26,31 +26,6 @@ class MetadataProcessor:
     def __init__(self, settings: Settings, translator: Translator):
         self.settings = settings
         self.translator = translator
-
-    def _parse_nfo_with_retry(self, nfo_path: Path) -> "ElementTree[ET.Element]":
-        """Parse NFO file with retry logic for incomplete/corrupt files.
-
-        Args:
-            nfo_path: Path to .nfo file to parse
-
-        Returns:
-            Parsed XML tree
-
-        Raises:
-            ET.ParseError: If file remains corrupt after retries
-            OSError: If file cannot be accessed after retries
-        """
-
-        @retry(
-            timeout=3.0,
-            interval=0.5,
-            log_interval=3.0,
-            exceptions=(ET.ParseError, OSError),
-        )
-        def parse_file() -> "ElementTree[ET.Element]":
-            return ET.parse(nfo_path)
-
-        return parse_file()
 
     def process_file(self, nfo_path: Path) -> MetadataProcessResult:
         """Process a single .nfo file with complete translation workflow.
@@ -65,7 +40,7 @@ class MetadataProcessor:
         metadata_info = None
         try:
             # Extract all metadata in single parse (including content for comparison)
-            metadata_info = self._extract_metadata_info(nfo_path)
+            metadata_info = extract_metadata_info(nfo_path)
 
             # Build TMDB IDs from parsed metadata using hierarchical resolution
             tmdb_ids = self._build_tmdb_ids_from_metadata(metadata_info, nfo_path)
@@ -199,60 +174,6 @@ class MetadataProcessor:
                 translated_content=None,
             )
 
-    def _extract_metadata_info(self, nfo_path: Path) -> MetadataInfo:
-        """Extract all metadata information from NFO file in single parse.
-
-        Args:
-            nfo_path: Path to .nfo file
-
-        Returns:
-            MetadataInfo object with all extracted data
-        """
-        tree = self._parse_nfo_with_retry(nfo_path)
-        root = tree.getroot()
-
-        # Determine file type from root tag
-        file_type = root.tag if root.tag in ("tvshow", "episodedetails") else "unknown"
-
-        info = MetadataInfo(file_type=file_type, xml_tree=tree)  # type: ignore[arg-type]
-
-        # Extract all uniqueid elements
-        for uniqueid in root.findall(".//uniqueid"):
-            id_type = uniqueid.get("type", "").lower()
-            id_value = uniqueid.text
-
-            if not id_value or not id_value.strip():
-                continue
-
-            if id_type == "tmdb":
-                info.tmdb_id = int(id_value.strip())
-            elif id_type == "tvdb":
-                info.tvdb_id = int(id_value.strip())
-            elif id_type == "imdb":
-                info.imdb_id = id_value.strip()
-
-        # Extract title
-        title_element = root.find("title")
-        if title_element is not None and title_element.text:
-            info.title = title_element.text.strip()
-
-        # Extract plot/description
-        plot_element = root.find("plot")
-        if plot_element is not None and plot_element.text:
-            info.description = plot_element.text.strip()
-
-        # For episode files, extract season/episode numbers
-        if file_type == "episodedetails":
-            season_element = root.find("season")
-            episode_element = root.find("episode")
-
-            if season_element is not None and season_element.text:
-                info.season = int(season_element.text.strip())
-            if episode_element is not None and episode_element.text:
-                info.episode = int(episode_element.text.strip())
-
-        return info
-
     def _resolve_tmdb_id_with_metadata(
         self, metadata_info: MetadataInfo, nfo_path: Path
     ) -> int | None:
@@ -296,7 +217,7 @@ class MetadataProcessor:
             if tvshow_path.exists() and tvshow_path.is_file():
                 try:
                     # Parse and extract metadata info
-                    metadata_info = self._extract_metadata_info(tvshow_path)
+                    metadata_info = extract_metadata_info(tvshow_path)
                     if metadata_info.file_type == "tvshow":
                         return metadata_info
                 except (ET.ParseError, ValueError, AttributeError):
@@ -633,7 +554,7 @@ class MetadataProcessor:
             return None
 
         try:
-            return self._extract_metadata_info(backup_path)
+            return extract_metadata_info(backup_path)
         except Exception as e:
             logger.warning(
                 f"Failed to read backup file {backup_path}: {e}", exc_info=True
